+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- python -m bin bumpversion changes the CHANGELOG.md for a release changelog
+
 ### Fixed
 - Links to versions in CHANGELOG.md